task 10

+++ 0x01-python-if_else_loops_functions/11-pow.py
@@ -3,4 +3,4 @@ def pow(a, b):
     if b >= 0:
         return int(a**b)
     else:
-        return int(1/(a**abs(b)))
+        return (1/(a**abs(b)))
